Added detail tables for displaying Factors and Variates. Resets all internal selections(studies, datasets, factors and variates) when database location has been changed. Resets all internal selections affected when dataset detail has been changed. Resets internal selections when exiting the modal window for viewing studies and datasets. Added basic notifications when an entry must be selected. Dataset export function integrated and functioning. Added dependency to apache commons IO 2.4 for File manipulation. Work has been tested and optimized.
Work left for this story: refinement of the overall look - some changes on the layout percentages perhaps, renaming of some texts as needed, some internationalizations, bold fonts for general headers etc. Redirect export path to the one described on the story description, just need a bit of clarification and its done.